Putting the terminal in transmit mode
This adds some additional DevFns for switching into and out of keypad
transmit mode. While the terminal is in keypad transmit mode, it sends
different escape sequences when arrow keys are pressed. These are the
sequences that correspond to terminfo entries, so the terminal must be
in keypad transmit mode to properly support arrow keys.